Remove dependence on utils and config
