+++ api/utils/cfApiClient.js
@@ -1,6 +1,55 @@
 const CloudFoundryAuthClient = require('./cfAuthClient');
 const HttpClient = require('./httpClient');
-const { filterEntity, firstEntity, objToQueryParams } = require('.');
+
+function filterEntity(res, name, field = 'name') {
+  const errMsg = `Not found: Entity @${field} = ${name}`;
+  const filtered = res.resources.filter(item => item.entity[field] === name);
+  if (filtered.length === 0) {
+    const error = new Error(errMsg);
+    error.name = name;
+    throw error;
+  }
+  return filtered;
+}
+
+function findEntity(res, name, field) {
+  return filterEntity(res, name, field)[0];
+}
+
+function findS3ServicePlan(res, name, s3ServicePlanId) {
+  const filtered = filterEntity(res, name);
+
+  if (name === 'basic-public') {
+    const servicePlan = filtered.find(f => f.entity.unique_id === s3ServicePlanId);
+    if (!servicePlan) {
+      const error = new Error(`Not found: @basic-public service plan = (${s3ServicePlanId})`);
+      error.name = name;
+      throw error;
+    }
+    return servicePlan;
+  }
+  return filtered[0];
+}
+
+function firstEntity(res, name) {
+  if (res.resources.length === 0) {
+    const error = new Error('Not found');
+    error.name = name;
+    throw error;
+  }
+
+  return res.resources[0];
+}
+
+function objToQueryParams(obj) {
+  const qs = new URLSearchParams();
+  Object
+    .entries(obj)
+    .forEach(([key, value]) => {
+      qs.append(key, value);
+    });
+  return qs;
+}
 
 class CloudFoundryAPIClient {
   constructor({ apiUrl, authClient } = {}) {
@@ -99,8 +148,8 @@ class CloudFoundryAPIClient {
     ));
   }
 
-  createS3ServiceInstance(name, serviceName, spaceGuid) {
-    return this.fetchS3ServicePlanGUID(serviceName)
+  createS3ServiceInstance(name, serviceName, spaceGuid, s3ServicePlanId) {
+    return this.fetchS3ServicePlanGUID(serviceName, s3ServicePlanId)
       .then((servicePlanGuid) => {
         const body = {
           name,
@@ -131,8 +180,8 @@ class CloudFoundryAPIClient {
     ));
   }
 
-  createSiteBucket(name, spaceGuid, keyIdentifier = 'key', serviceName = 'basic-public') {
-    return this.createS3ServiceInstance(name, serviceName, spaceGuid)
+  createSiteBucket(name, spaceGuid, s3ServicePlanId, keyIdentifier = 'key', serviceName = 'basic-public') {
+    return this.createS3ServiceInstance(name, serviceName, spaceGuid, s3ServicePlanId)
       .then(res => this.createServiceKey(name, res.metadata.guid, keyIdentifier));
   }
 
@@ -148,7 +197,7 @@ class CloudFoundryAPIClient {
         `/v2/routes?q=host:${host}`,
         token
       ))
-      .then(res => filterEntity(res, host, 'host'))
+      .then(res => findEntity(res, host, 'host'))
       .then(entity => this.accessToken()
         .then(token => this.request(
           'DELETE',
@@ -172,7 +221,7 @@ class CloudFoundryAPIClient {
 
   fetchServiceInstance(name) {
     return this.fetchServiceInstances(name)
-      .then(res => filterEntity(res, name));
+      .then(res => findEntity(res, name));
   }
 
   fetchServiceInstanceCredentials(name) {
@@ -198,7 +247,7 @@ class CloudFoundryAPIClient {
 
   fetchServiceKey(name) {
     return this.fetchServiceKeys()
-      .then(res => filterEntity(res, name))
+      .then(res => findEntity(res, name))
       .then(key => this.accessToken().then(token => this.request(
         'GET',
         `/v2/service_keys/${key.metadata.guid}`,
@@ -214,12 +263,12 @@ class CloudFoundryAPIClient {
     ));
   }
 
-  fetchS3ServicePlanGUID(serviceName) {
+  fetchS3ServicePlanGUID(serviceName, s3ServicePlanId) {
     return this.accessToken().then(token => this.request(
       'GET',
       '/v2/service_plans',
       token
-    )).then(res => filterEntity(res, serviceName))
+    )).then(res => findS3ServicePlan(res, serviceName, s3ServicePlanId))
       .then(service => service.metadata.guid);
   }
 
@@ -258,5 +307,11 @@ class CloudFoundryAPIClient {
     return this.accessToken().then(token => this.request(method, path, token, json));
   }
 }
+
+CloudFoundryAPIClient.filterEntity = filterEntity;
+CloudFoundryAPIClient.findEntity = findEntity;
+CloudFoundryAPIClient.findS3ServicePlan = findS3ServicePlan;
+CloudFoundryAPIClient.firstEntity = firstEntity;
+CloudFoundryAPIClient.objToQueryParams = objToQueryParams;
 
 module.exports = CloudFoundryAPIClient;
